Alterado nome do arquivo MoneyWriter.cs
